Fix isoformat parsing
